fix(codebuild-setup): changes to packaging structure; minor fixes in README
