--- amplify-codegen-ci-support/src/ci_cd_pipeline/requirements.txt
@@ -1,1 +1,0 @@
--e .
