--- amplify-codegen-ci-support/src/ci_cd_pipeline/stacks/__init__.py
@@ -1,9 +1,0 @@
-import os
-import sys
-
-sys.path.append(
-    os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
-        ".",
-    )
-)
