+++ ci-cd/ci_cd_resources/.gitignore
@@ -9,3 +9,4 @@ __pycache__
 .cdk.staging
 cdk.out
 .venv
+.idea
