+++ ci-cd/ci_cd_resources/requirements.txt
@@ -1,0 +1,3 @@
+aws_cdk.aws_codebuild
+aws-cdk.core
+boto3
